requirements_secure.txt
langchain==0.1.0
openai==1.12.0
chromadb==0.4.22
streamlit==1.31.0
tiktoken==0.5.2
python-dotenv==1.0.0

